FlaskApp/requirements.txt
boto3==1.26.59
botocore==1.29.59
cachelib==0.10.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Session==0.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
attrs==22.2.0
autopep8==2.0.1
blis==0.7.8
boto3==1.26.59
botocore==1.29.59
cachelib==0.10.1
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.6
distlib==0.3.5
filelock==3.8.0
Flask==2.2.2
Flask-Login==0.6.2
flask-ngrok==0.0.25
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
fonttools==4.38.0
future==0.18.3
greenlet==1.1.2
gym==0.26.2
gym-notices==0.0.8
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.6.2
murmurhash==1.0.8
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.23.2
opencv-python==4.7.0.68
packaging==21.3
pandas==1.4.3
pathy==0.6.2
pika==1.3.1
Pillow==9.4.0
platformdirs==2.5.2
preshed==3.0.7
protobuf==3.20.3
psycopg2==2.9.5
pycodestyle==2.10.0
pydantic==1.9.2
pyparsing==3.0.9
pyswarm==0.6
pyswarms==1.3.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
scipy==1.10.1
six==1.16.0
smart-open==5.2.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.40
sqlclient==21.4.24
srsly==2.4.4
thinc==8.1.0
tomli==2.0.1
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.14
virtualenv==20.16.3
wasabi==0.10.1
Werkzeug==2.2.2
